% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{fatal_error}
\alias{fatal_error}
\title{Report a fatal error and terminate execution}
\usage{
fatal_error(message, call = parent.frame())
}
\arguments{
\item{message}{the character vector containing the error message,
can be formatted in the style of \href{https://rlang.r-lib.org/reference/format_error_bullets.html}{rlang::format_error_bullets}}

\item{call}{either a function call or a call frame in which the error
has occured}
}
\description{
\code{fatal_error()} is similar to \link[base:stop]{base::stop}, except it bypasses
the error handling mechanism and will always immediately terminate
the program. Use this function to report critical execution errors,
where recovery is impossible or unfeasible.
}
\details{
In the interactive mode, \code{fatal_error()} will exit to the interpreter prompt.
In a script, it will terminate execution using \link[base:quit]{base::quit}

\code{fatal_error()} is used by \link{sanity_check} to immediately terminate execution
when a critical assertion in program logic has failed.
}
